refactor: split into files
Split the `x509/` package into smaller files, previously it was 1500
lines monolith with different methods and types mixed together, split it
so it's mostly one type per file (or closely related files). I added
a couple of unit-tests, but no code changes.
