+++ .golangci.yml
@@ -1,117 +1,32 @@
 # THIS FILE WAS AUTOMATICALLY GENERATED, PLEASE DO NOT EDIT.
 #
-# Generated on 2024-12-04T08:15:33Z by kres 232fe63.
+# Generated on 2025-05-14T13:00:06Z by kres 5ad3e5f.
+
+version: "2"
 
 # options for analysis running
 run:
-  timeout: 10m
+  modules-download-mode: readonly
   issues-exit-code: 1
   tests: true
-  build-tags: [ ]
-  modules-download-mode: readonly
 
 # output configuration options
 output:
   formats:
-    - format: colored-line-number
+    text:
       path: stdout
-  print-issued-lines: true
-  print-linter-name: true
-  uniq-by-line: true
+      print-issued-lines: true
+      print-linter-name: true
   path-prefix: ""
 
-# all available settings of specific linters
-linters-settings:
-  dogsled:
-    max-blank-identifiers: 2
-  dupl:
-    threshold: 150
-  errcheck:
-    check-type-assertions: true
-    check-blank: true
-  exhaustive:
-    default-signifies-exhaustive: false
-  gci:
-    sections:
-      - standard # Standard section: captures all standard packages.
-      - default # Default section: contains all imports that could not be matched to another section type.
-      - localmodule # Imports from the same module.
-  gocognit:
-    min-complexity: 30
-  nestif:
-    min-complexity: 5
-  goconst:
-    min-len: 3
-    min-occurrences: 3
-  gocritic:
-    disabled-checks: [ ]
-  gocyclo:
-    min-complexity: 20
-  godot:
-    scope: declarations
-  gofmt:
-    simplify: true
-  gomodguard: { }
-  govet:
-    enable-all: true
-  lll:
-    line-length: 200
-    tab-width: 4
-  misspell:
-    locale: US
-    ignore-words: [ ]
-  nakedret:
-    max-func-lines: 30
-  prealloc:
-    simple: true
-    range-loops: true # Report preallocation suggestions on range loops, true by default
-    for-loops: false # Report preallocation suggestions on for loops, false by default
-  nolintlint:
-    allow-unused: false
-    allow-no-explanation: [ ]
-    require-explanation: false
-    require-specific: true
-  rowserrcheck: { }
-  testpackage: { }
-  unparam:
-    check-exported: false
-  unused:
-    local-variables-are-used: false
-  whitespace:
-    multi-if: false   # Enforces newlines (or comments) after every multi-line if statement
-    multi-func: false # Enforces newlines (or comments) after every multi-line function signature
-  wsl:
-    strict-append: true
-    allow-assign-and-call: true
-    allow-multiline-assign: true
-    allow-cuddle-declarations: false
-    allow-trailing-comment: false
-    force-case-trailing-whitespace: 0
-    force-err-cuddling: false
-    allow-separated-leading-comment: false
-  gofumpt:
-    extra-rules: false
-  cyclop:
-    # the maximal code complexity to report
-    max-complexity: 20
-  depguard:
-    rules:
-      prevent_unmaintained_packages:
-        list-mode: lax # allow unless explicitly denied
-        files:
-          - $all
-        deny:
-          - pkg: io/ioutil
-            desc: "replaced by io and os packages since Go 1.16: https://tip.golang.org/doc/go1.16#ioutil"
 
 linters:
-  enable-all: true
-  disable-all: false
-  fast: false
+  default: all
   disable:
     - exhaustruct
     - err113
     - forbidigo
+    - funcorder
     - funlen
     - gochecknoglobals
     - gochecknoinits
@@ -132,18 +47,109 @@ linters:
     - testifylint # complains about our assert recorder and has a number of false positives for assert.Greater(t, thing, 1)
     - protogetter # complains about us using Value field on typed spec, instead of GetValue which has a different signature
     - perfsprint # complains about us using fmt.Sprintf in non-performance critical code, updating just kres took too long
-    - goimports # same as gci
     - musttag # seems to be broken - goes into imported libraries and reports issues there
+    - nolintlint # gives false positives - disable until https://github.com/golangci/golangci-lint/issues/3228 is resolved
+  # all available settings of specific linters
+  settings:
+    cyclop:
+      # the maximal code complexity to report
+      max-complexity: 20
+    dogsled:
+      max-blank-identifiers: 2
+    dupl:
+      threshold: 150
+    errcheck:
+      check-type-assertions: true
+      check-blank: true
+    exhaustive:
+      default-signifies-exhaustive: false
+    gocognit:
+      min-complexity: 30
+    nestif:
+      min-complexity: 5
+    goconst:
+      min-len: 3
+      min-occurrences: 3
+    gocritic:
+      disabled-checks: [ ]
+    gocyclo:
+      min-complexity: 20
+    godot:
+      scope: declarations
+    gomodguard: { }
+    govet:
+      enable-all: true
+    lll:
+      line-length: 200
+      tab-width: 4
+    misspell:
+      locale: US
+    nakedret:
+      max-func-lines: 30
+    prealloc:
+      simple: true
+      range-loops: true # Report preallocation suggestions on range loops, true by default
+      for-loops: false # Report preallocation suggestions on for loops, false by default
+    rowserrcheck: { }
+    testpackage: { }
+    unparam:
+      check-exported: false
+    unused:
+      local-variables-are-used: false
+    whitespace:
+      multi-if: false   # Enforces newlines (or comments) after every multi-line if statement
+      multi-func: false # Enforces newlines (or comments) after every multi-line function signature
+    wsl:
+      strict-append: true
+      allow-assign-and-call: true
+      allow-multiline-assign: true
+      allow-trailing-comment: false
+      force-case-trailing-whitespace: 0
+      allow-separated-leading-comment: false
+      allow-cuddle-declarations: false
+      force-err-cuddling: false
+    depguard:
+      rules:
+        prevent_unmaintained_packages:
+          list-mode: lax # allow unless explicitly denied
+          files:
+            - $all
+          deny:
+            - pkg: io/ioutil
+              desc: "replaced by io and os packages since Go 1.16: https://tip.golang.org/doc/go1.16#ioutil"
 
+  exclusions:
+    generated: lax
+    paths:
+      - third_party$
+      - builtin$
+      - examples$
 issues:
-  exclude: [ ]
-  exclude-rules: [ ]
-  exclude-use-default: false
-  exclude-case-sensitive: false
   max-issues-per-linter: 10
   max-same-issues: 3
+  uniq-by-line: true
   new: false
 
 severity:
-  default-severity: error
-  case-sensitive: false
+  default: error
+formatters:
+  enable:
+    - gci
+    - gofmt
+    - gofumpt
+  settings:
+    gci:
+      sections:
+        - standard
+        - default
+        - localmodule
+    gofmt:
+      simplify: true
+    gofumpt:
+      extra-rules: false
+  exclusions:
+    generated: lax
+    paths:
+      - third_party$
+      - builtin$
+      - examples$
